eduhelpify/pipelines/process_docs/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
asttokens==3.0.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comm==0.2.2
debugpy==1.8.14
decorator==5.2.1
deprecation==2.1.0
docutils==0.21.2
exceptiongroup==1.2.2
executing==2.2.0
Flask==2.3.3
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gotrue==1.3.1
grpcio==1.71.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==0.17.3
httplib2==0.22.0
httpx==0.24.1
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.29.5
ipython==8.18.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.4.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
nest-asyncio==1.6.0
packaging==25.0
parso==0.8.4
pillow==11.2.1
platformdirs==4.3.7
postgrest==0.13.2
prompt_toolkit==3.0.51
proto-plus==1.26.1
protobuf==4.25.7
psutil==7.0.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.4
pydantic_core==2.33.2
Pygments==2.19.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.0
python-http-client==3.3.7
python-pptx==1.0.2
pywin32==310
pyzmq==26.4.0
realtime==1.0.6
reportlab==4.4.0
requests==2.32.3
rsa==4.9.1

six==1.17.0
sniffio==1.3.1
stack-data==0.6.3
starkbank-ecdsa==2.2.0
storage3==0.6.1
StrEnum==0.4.15
supabase==2.0.3
supafunc==0.3.3
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
uritemplate==4.1.1
urllib3==2.4.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.1.3
XlsxWriter==3.2.3
zipp==3.21.0
